Address issue with byte strings that are not null terminated
This fixes a bug where properties with signature "ay" had to be null
terminated, which is not always the case.

The issue is that Glib::Variant<std::string>::get() uses
g_variant_get_bytestring() which will return an empty string if the
array does not end with null. It will also only return a part of the
array if the array contains a null character before the end of the
array.

Special handling of signature "ay" is added, so it will use
g_variant_get_fixed_array() instead which works well with any kind of
array payload.